@@ -29,11 +29,10 @@ import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
-import android.graphics.Color;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v4.os.ConfigurationCompat;
+import androidx.annotation.NonNull;
+import androidx.core.os.ConfigurationCompat;
 import android.util.Log;
 
 import com.google.android.gms.tasks.OnCompleteListener;
